The Player class

The Player class

Performance

The features of the agents are defined within class Player: the prior Gaussian distribution characterized by the mean (mu) and the standard deviation (sigma), the standard deviation of the performance (beta), and the dynamic uncertainty of the skill (gamma).

TrueSkillThroughTime.Player — Type

The Player class is used to define the features of the agents. We can create objects by indicating the parameters in order or by mentioning their names.

Player(prior::Gaussian=Gaussian(MU,SIGMA), beta::Float64=BETA, gamma::Float64=GAMMA)
Player(;prior::Gaussian=Gaussian(MU,SIGMA), beta::Float64=BETA, gamma::Float64=GAMMA)

prior is the prior belief distribution of skill hypotheses

beta is the standar deviation of the agent performance

gamma is the uncertainty (standar deviation) added to the estimates as time progresses

The default value of MU, SIGMA, BETA and GAMMA are

julia> a1 = ttt.Player()Player(Gaussian(mu=0.0, sigma=6.0), beta=1.0, gamma=0.03)

julia> a2 = ttt.Player(ttt.Gaussian(0.0, 1.0))Player(Gaussian(mu=0.0, sigma=1.0), beta=1.0, gamma=0.03)

We can also create special players who have non-random performances (beta=0.0), and whose skills do not change over time (gamma=0.0).

julia> a3 = ttt.Player(beta=0.0, gamma=0.0)Player(Gaussian(mu=0.0, sigma=6.0), beta=0.0, gamma=0.0)
julia> a3.beta0.0
julia> a3.gamma0.0

Performance

The performances $p$ are random variables around their unknown true skill $s$,

$p \sim \mathcal{N}(s,\beta^2)$

TrueSkillThroughTime.performance — Method

performance(R::Player)

julia> ttt.performance(a2)Gaussian(mu=0.0, sigma=1.414214)
julia> ttt.performance(a3)Gaussian(mu=0.0, sigma=6.0)